requirements.txt
streamlit
python-dotenv
matplotlib
pandas
numpy
seaborn
openai
langchain
langchainhub
langchain-openai
langchain-anthropic
langchain-google-genai
langchain_experimental
langchain_cohere
tabulate
black
isort

